fixed GET all students by teacher route
