+++ src/controllers/teacher-controller.ts
@@ -2,6 +2,7 @@ import { Request, Response } from "express";
 export interface TeacherController {
   createTeacher(req: Request, res: Response): Promise<void>;
   getAllTeachers(req: Request, res: Response): Promise<void>;
+  getAllStudentsByTeacher(req: Request, res: Response): Promise<void>;
   editTeacher(req: Request, res: Response): Promise<void>;
   deleteTeacher(req: Request, res: Response): Promise<void>;
   getTeacherByLogin(req: Request, res:Response): Promise<void>;
