+++ src/repositories/impls/student-repository-impl.ts
@@ -25,7 +25,8 @@ export class StudentRepositoryImpl implements StudentRepository {
       studentBody.degreeLevelId,
       studentBody.course,
       studentBody.admissionYear,
-      studentBody.socials)
+      studentBody.socials,
+      studentBody.teacherLogin)
     );
 
     const results = await this.transactionRunner.run(queryConstructors);
@@ -43,7 +44,7 @@ export class StudentRepositoryImpl implements StudentRepository {
       course: studentData.course,
       admissionYear: studentData.admission_year,
       socials: studentData.socials,
-      teacherLogin: studentData.teacherLogin
+      teacherLogin: studentData.teacher_login
     };
   };
 
@@ -67,35 +68,7 @@ export class StudentRepositoryImpl implements StudentRepository {
       course: data.course,
       admission_year: data.admission_year,
       socials: data.socials,
-      teacherLogin: studentData.teacherLogin
-    }));
-
-    return studentModels;
-  };
-
-
-
-  public async getAllStudentsByTeacher(teacherLogin: string): Promise<Array<StudentModel>> {
-    const queryConstructors: Array<SingleQueryConstructor> = new Array<SingleQueryConstructor>();
-
-    queryConstructors.push(this.studentQueries.getAllStudentsByTeacher(teacherLogin));
-
-    const results = await this.transactionRunner.run(queryConstructors);
-
-    Assert.notNullOrUndefined(results, `There are no students for the teacher ${teacherLogin} in the database`);
-
-    const studentData = results[0];
-
-    const studentModels = studentData.map((data: any) => ({
-      login: data.login,
-      fullName: data.full_name,
-      phoneNumber: data.phone_number,
-      studyProgramId: data.study_program_id,
-      degreeLevelId: data.degree_level_id,
-      course: data.course,
-      admission_year: data.admission_year,
-      socials: data.socials,
-      teacherLogin: studentData.teacherLogin
+      teacherLogin: data.teacher_login
     }));
 
     return studentModels;
@@ -134,7 +107,7 @@ export class StudentRepositoryImpl implements StudentRepository {
       course: studentData.course,
       admissionYear: studentData.admission_year,
       socials: studentData.socials,
-      teacherLogin: studentData.teacherLogin
+      teacherLogin: studentData.teacher_login
     };
   };
 
@@ -158,7 +131,7 @@ export class StudentRepositoryImpl implements StudentRepository {
       course: studentData.course,
       admissionYear: studentData.admission_year,
       socials: studentData.socials,
-      teacherLogin: studentData.teacherLogin
+      teacherLogin: studentData.teacher_login
     };
   };
 
@@ -186,7 +159,7 @@ export class StudentRepositoryImpl implements StudentRepository {
       course: studentData.course,
       admissionYear: studentData.admission_year,
       socials: studentData.socials,
-      teacherLogin: studentData.teacherLogin
+      teacherLogin: studentData.teacher_login
     }
   }
